requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-decouple
psycopg2-binary
httpx
firebase-admin
slowapi
python-multipart
gunicorn
reportlab
Pillow